AudioContext stops playing when minimizing or moving the macOS Safari window to the background. https://bugs.webkit.org/show_bug.cgi?id=231105 <rdar://problem/83889697>
Reviewed by Eric Carlson.

Source/WebCore:

AudioContext was stopped when minimizing windows if not playing audio as a way to reduce CPU usage.
But this is visible to users (if AudioContext produces audio through a track) or to web pages (use of audio worklet or script processor node).
Remove that code.
Small refactoring to remove isOfflineContext checks that can only be false in AudioContext.

Covered by existing tests.

* Modules/webaudio/AudioContext.cpp:
* Modules/webaudio/AudioContext.h:

LayoutTests:

Use internals API to be working on iOS as well (this does not trigger enterBackground code path).

* webaudio/silent-audio-interrupted-in-background-expected.txt:
* webaudio/silent-audio-interrupted-in-background.html:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291267 268f45cc-cd09-0410-ab3c-d52691b4dbfc