[webkitpy] Make SimulatedDevice.is_usable check more robust https://bugs.webkit.org/show_bug.cgi?id=237795
Reviewed by Jonathan Bedard.

Rather than searching for the service name in the output of `launchctl list`,
pass in the name as an additional argument so that it prints out information
about that particular service. I have verified that this works in local testing
for both iOS and watchOS simulators.

* Scripts/webkitpy/xcode/simulated_device.py:
(SimulatedDevice):
(SimulatedDevice.is_usable):

Canonical link: https://commits.webkit.org/248439@main

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291300 268f45cc-cd09-0410-ab3c-d52691b4dbfc