Stop using DYLD_INTERPOSE https://bugs.webkit.org/show_bug.cgi?id=237867
Patch by Alex Christensen <achristensen@webkit.org> on 2022-03-15
Reviewed by Anders Carlsson.

Source/WebCore:

It never worked on iOS, and it doesn't work on M1 Macs.
It used to be used for NPAPI and the SecItemShim,
but those have been removed and updated to use other methods that work on M1 Macs.
It is currently only used for tests in editing/secure-input which continue to pass
as much as they did before this change.

* Configurations/WebCoreTestShim.xcconfig: Removed.
* WebCore.xcodeproj/project.pbxproj:
* platform/mac/DynamicLinkerInterposing.h: Removed.
* testing/WebCoreTestShimLibrary.cpp: Removed.

Tools:

* Scripts/webkitpy/port/mac.py:
(MacPort.setup_environ_for_server):
* Scripts/webkitpy/port/mac_unittest.py:
(MacTest.test_setup_environ_for_server):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291302 268f45cc-cd09-0410-ab3c-d52691b4dbfc