[GTK][WPE] Guard libgbm and libdrm code, make them conditional dependencies for GTK https://bugs.webkit.org/show_bug.cgi?id=237974
Patch by Zan Dobersek <zdobersek@igalia.com> on 2022-03-17
Reviewed by Alejandro G. Castro.

.:

Guard usage of libgbm and libdrm dependencies with the USE_LIBGBM
define. This should be required until we can switch to a possibly more
dynamic implementation that doesn't require build-time dependencies.

For the GTK port, the desire is to keep these dependencies fully
optional. At the moment they are required when enabling ANGLE-backed
WebGL, with the USE_LIBGBM also defined in that case. This is needed
because ANGLE integration for Linux ports already depends on these
libraries.

For the WPE port only USE_LIBGBM is defined and the two dependencies
remain required.

* Source/cmake/OptionsGTK.cmake:
* Source/cmake/OptionsWPE.cmake:

Source/WebCore:

Add USE(LIBGBM) guards around GBMDevice and GBMBufferSwapchain code,
and compile against the libgbm and libdrm dependencies only when
USE_LIBGBM is defined. This makes the two dependencies optional if
desired.

* PlatformGTK.cmake:
* PlatformWPE.cmake:
* platform/graphics/gbm/GBMBufferSwapchain.cpp:
* platform/graphics/gbm/GBMBufferSwapchain.h:
* platform/graphics/gbm/GBMDevice.cpp:
* platform/graphics/gbm/GBMDevice.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291392 268f45cc-cd09-0410-ab3c-d52691b4dbfc