REGRESSION (iOS 15.4 / r287669): Mobile app stopped working due to CSS / angular animation https://bugs.webkit.org/show_bug.cgi?id=237920 LayoutTests/imported/w3c:
<rdar://problem/90335275>

Reviewed by Antti Koivisto.

Add a test that looks at the default computed style for the "transition" and "animation"
shorthands.

* web-platform-tests/css/css-animations/parsing/animation-computed-expected.txt:
* web-platform-tests/css/css-animations/parsing/animation-computed.html:
* web-platform-tests/css/css-pseudo/first-letter-allowed-properties-expected.txt:
* web-platform-tests/css/css-pseudo/first-line-allowed-properties-expected.txt:
* web-platform-tests/css/css-transitions/parsing/transition-computed-expected.txt:
* web-platform-tests/css/css-transitions/parsing/transition-computed.html:

Source/WebCore:

Reviewed by Antti Koivisto.

We refactored the computed style output for the "transition" and "animation" shorthand
properties in r287669 and always returned a CSSValueList, even when the list of transitions
or animations was empty. This meant that for an element without any transition or animation
properties, we'd return the empty string for those properties.

Now we return the computed style for a default animation instead.

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::animationShorthandValue):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291420 268f45cc-cd09-0410-ab3c-d52691b4dbfc