Avoid calling connectionClosedFromServer on a connection twice when a version change is active https://bugs.webkit.org/show_bug.cgi?id=215271
Reviewed by Sihui Liu.

Source/WebCore:

Since bug 203690, all IDB work is done on a single background thread.
A consequence of this change is that there are states when
UniqueIDBDatabase::m_versionChangeDatabaseConnection is also in
m_openDatabaseConnections. We should avoid calling
connectionClosedFromServer on the same connection twice, since that
can result in a spurious second error event being sent to the
IDBDatabase JS object if connections are closed while a version chnge
is in progress.

* Modules/indexeddb/server/UniqueIDBDatabase.cpp:
(WebCore::IDBServer::UniqueIDBDatabase::immediateClose):

LayoutTests:

* platform/ios-wk2/TestExpectations:
* platform/mac/TestExpectations:
* platform/win/TestExpectations:
* platform/wincairo/TestExpectations:
Remove test flakiness annotations.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291435 268f45cc-cd09-0410-ab3c-d52691b4dbfc