REGRESSION (r290628): Scrubber makes a visual trail when scrubbing on tv.youtube.com https://bugs.webkit.org/show_bug.cgi?id=238047 <rdar://problem/90451319>
Reviewed by Alan Bujtas.

Source/WebCore:

In r290628 I added a call to clearRepaintRects() when changing the backing sharing
target layer for a RenderLayer. However, for a layer using backing sharing, this
gets called twice during a normal compositing update, which has the effect of
clearing the cached repaint rects so that a layer repaint always computes new rects.
Thus, if a layer moved, we'd fail to repaint the old rect.

RenderLayerBacking::setBackingSharingLayers() already has logic to update repaint
rects when state changes, so we don't need this clearing.

Test: compositing/repaint/move-backing-sharing-layer.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::setBackingProviderLayer):

LayoutTests:

* compositing/repaint/move-backing-sharing-layer-expected.txt: Added.
* compositing/repaint/move-backing-sharing-layer.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291493 268f45cc-cd09-0410-ab3c-d52691b4dbfc