Web Inspector: `console.screenshot` has extra transparent pixels at bottom of screenshot on platforms with viewports that are obscured by browser chrome https://bugs.webkit.org/show_bug.cgi?id=238060
Reviewed by Devin Rousso.

ScrollView::sizeForVisibleRect includes partially-visible portions of the viewport, including in Safari's case
the part of the viewport under the semi-translucent title bar. Because we force an origin of zero, this meant
that we were overshooting the bottom of the viewport by the height of this area when taking a snapshot for
`console.screenshot()`. Instead, we should use unobscured viewport rectangle and provide coordinates in
document-space, not view-space to `WebCore::snapshotFrameRect`.

* page/PageConsoleClient.cpp:
(WebCore::PageConsoleClient::screenshot):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@291519 268f45cc-cd09-0410-ab3c-d52691b4dbfc